refactor: mapper 객체를 config/constants 세그먼트로 이동
