+++ apps/web/src/entities/ceremony/config/constants/index.ts
@@ -1,0 +1,1 @@
+export { MY_CEREMONY_STATE_MAP } from './my-ceremony-state-map';
