+++ apps/web/src/entities/ceremony/config/index.ts
@@ -1,1 +1,2 @@
 export { getCeremonyIcon } from './ceremonyIcon';
+export { MY_CEREMONY_STATE_MAP } from './constants';
