+++ apps/web/src/entities/ceremony/model/types.ts
@@ -47,16 +47,6 @@ export type CeremonyFilterType = '전체' | '경사' | '조사';
 /** 내 경조사 상태 필터 */
 export type MyCeremonyStateFilter = '등록 완료' | '등록 거부' | '등록 대기중';
 
-/** 내 경조사 상태 필터 → CeremonyState 매핑 */
-export const MY_CEREMONY_STATE_MAP: Record<
-  MyCeremonyStateFilter,
-  CeremonyState
-> = {
-  '등록 완료': 'ACCEPT',
-  '등록 거부': 'REJECT',
-  '등록 대기중': 'AWAIT',
-};
-
 export type CeremonyTypeApi = 'CELEBRATION' | 'CONDOLENCE';
 
 export type CeremonyCategoryApi = 'ETC' | string;
